+++ extensions/github-review-requests/src/util.ts
@@ -1,0 +1,10 @@
+import TimeAgo from "javascript-time-ago";
+import en from "javascript-time-ago/locale/en.json";
+
+TimeAgo.addDefaultLocale(en);
+const timeAgo = new TimeAgo("en-US");
+
+export function formatDate(input: Date | string) {
+  const date = typeof input === "string" ? new Date(input) : input;
+  return timeAgo.format(date, "twitter") as string;
+}