+++ extensions/pokedex/src/components/detail.tsx
@@ -1,4 +1,4 @@
-import { Action, ActionPanel, Detail, getPreferenceValues } from "@raycast/api";
+import { Color, Detail, getPreferenceValues } from "@raycast/api";
 import { useEffect, useMemo, useState } from "react";
 import json2md from "json2md";
 import { getPokemon } from "../api";
@@ -22,6 +22,27 @@ enum GrowthRate {
   "Erratic" = 5,
   "Fluctuating" = 6,
 }
+
+const typeColor: { [key: string]: string } = {
+  Normal: "#a8a77a",
+  Fire: "#ee8130",
+  Water: "#6390f0",
+  Electric: "#f7d02c",
+  Grass: "#7ac74c",
+  Ice: "#96d9d6",
+  Fighting: "#c22e28",
+  Poison: "#a33ea1",
+  Ground: "#e2bf65",
+  Flying: "#a98ff3",
+  Psychic: "#f95587",
+  Bug: "#a6b91a",
+  Rock: "#b6a136",
+  Ghost: "#735797",
+  Dragon: "#6f35fc",
+  Dark: "#705746",
+  Steel: "#b7b7ce",
+  Fairy: "#d685ad",
+};
 
 function random(lower: number, upper: number) {
   return lower + Math.floor(Math.random() * (upper - lower + 1));
@@ -85,12 +106,7 @@ export default function PokemonDetail(props: { id?: number }) {
   const dataObject: json2md.DataObject = useMemo(() => {
     if (!pokemon) return [];
 
-    const {
-      pokemon_v2_pokemonabilities,
-      pokemon_v2_pokemonspecy,
-      pokemon_v2_pokemontypes,
-      pokemon_v2_pokemonstats,
-    } = pokemon;
+    const { pokemon_v2_pokemonspecy } = pokemon;
 
     const {
       pokemon_v2_evolutionchain,
@@ -176,49 +192,6 @@ export default function PokemonDetail(props: { id?: number }) {
             source: `https://assets.pokemon.com/assets/cms2/img/pokedex/detail/${pkmNumber}.png`,
           },
         ],
-      },
-      {
-        h2: "Pokédex data",
-      },
-      {
-        p:
-          "_Type:_ " +
-          pokemon_v2_pokemontypes
-            .map((n) => n.pokemon_v2_type.pokemon_v2_typenames[0].name)
-            .join(", "),
-      },
-      { p: `_Height:_ ${pokemon.height / 10}m` },
-      { p: `_Weight:_ ${pokemon.weight / 10}kg` },
-      {
-        p: `_Abilities:_ ${pokemon_v2_pokemonabilities
-          .map((a) => {
-            if (a.is_hidden) {
-              return `${a.pokemon_v2_ability.pokemon_v2_abilitynames[0].name} (hidden)`;
-            }
-
-            return a.pokemon_v2_ability.pokemon_v2_abilitynames[0].name;
-          })
-          .join(", ")}`,
-      },
-      {
-        h2: "Base stats",
-      },
-      ...pokemon_v2_pokemonstats.map((n) => {
-        if (n.effort) {
-          ev.push(
-            `${n.effort} ${n.pokemon_v2_stat.pokemon_v2_statnames[0].name}`
-          );
-        }
-
-        return {
-          p: `_${n.pokemon_v2_stat.pokemon_v2_statnames[0].name}_: ${n.base_stat}`,
-        };
-      }),
-      {
-        p: `Total: **${pokemon_v2_pokemonstats.reduce(
-          (prev, cur) => prev + cur.base_stat,
-          0
-        )}**`,
       },
       {
         h2: "Training",
@@ -329,10 +302,7 @@ export default function PokemonDetail(props: { id?: number }) {
     return data;
   }, [pokemon]);
 
-  const englishName = () => {
-    // 9 is language_id for English
-    return nameByLang["9"].name.replace(/ /g, "_");
-  };
+  const englishName = nameByLang["9"]?.name.replace(/ /g, "_");
 
   return (
     <Detail
@@ -341,22 +311,63 @@ export default function PokemonDetail(props: { id?: number }) {
         pokemon ? `${nameByLang[language].name} | Pokédex` : "Pokédex"
       }
       markdown={json2md(dataObject)}
-      actions={
+      metadata={
         pokemon && (
-          <ActionPanel>
-            <ActionPanel.Section title="Pokémon">
-              <Action.OpenInBrowser
-                title="Open in the Official Pokémon Website"
-                icon="pokeball.svg"
-                url={`https://www.pokemon.com/us/pokedex/${pokemon.pokemon_v2_pokemonspecy.name}`}
-              />
-              <Action.OpenInBrowser
-                title="Open in Bulbapedia"
-                icon="bulbapedia.svg"
-                url={`https://bulbapedia.bulbagarden.net/wiki/${englishName()}_(Pok%C3%A9mon)`}
-              />
-            </ActionPanel.Section>
-          </ActionPanel>
+          <Detail.Metadata>
+            <Detail.Metadata.Link
+              title="Official Pokémon Website"
+              text={nameByLang[language].name}
+              target={`https://www.pokemon.com/us/pokedex/${pokemon.pokemon_v2_pokemonspecy.name}`}
+            />
+            <Detail.Metadata.Link
+              title="Bulbapedia"
+              text={englishName}
+              target={`https://bulbapedia.bulbagarden.net/wiki/${englishName}_(Pok%C3%A9mon)}`}
+            />
+            <Detail.Metadata.Separator />
+            <Detail.Metadata.Label
+              title="Height"
+              text={`${pokemon.height / 10}m`}
+            />
+            <Detail.Metadata.Label
+              title="Weight"
+              text={`${pokemon.weight / 10}kg`}
+            />
+            <Detail.Metadata.TagList title="Type">
+              {pokemon.pokemon_v2_pokemontypes.map((t) => {
+                return (
+                  <Detail.Metadata.TagList.Item
+                    key={t.pokemon_v2_type.pokemon_v2_typenames[0].name}
+                    text={t.pokemon_v2_type.pokemon_v2_typenames[0].name}
+                    color={
+                      typeColor[t.pokemon_v2_type.pokemon_v2_typenames[0].name]
+                    }
+                  />
+                );
+              })}
+            </Detail.Metadata.TagList>
+            <Detail.Metadata.TagList title="Abilities">
+              {pokemon.pokemon_v2_pokemonabilities.map((t) => {
+                return (
+                  <Detail.Metadata.TagList.Item
+                    key={t.pokemon_v2_ability.pokemon_v2_abilitynames[0].name}
+                    text={t.pokemon_v2_ability.pokemon_v2_abilitynames[0].name}
+                    color={t.is_hidden ? Color.Brown : Color.Blue}
+                  />
+                );
+              })}
+            </Detail.Metadata.TagList>
+            <Detail.Metadata.Separator />
+            {pokemon.pokemon_v2_pokemonstats.map((stat, idx) => {
+              return (
+                <Detail.Metadata.Label
+                  key={idx}
+                  title={stat.pokemon_v2_stat.pokemon_v2_statnames[0].name}
+                  text={stat.base_stat.toString()}
+                />
+              );
+            })}
+          </Detail.Metadata>
         )
       }
     />