+++ extensions/vim-bro/src/index.tsx
@@ -1,6 +1,7 @@
 import { List } from "@raycast/api";
 import { useEffect, useState } from "react";
 import commandsRaw from "./commands.json";
+<<<<<<< HEAD
 
 type Command = {
   kbd: string;
@@ -11,6 +12,10 @@ type CommandGroup = {
   key: string;
   commands: Command[];
 };
+=======
+import { Command, CommandGroup } from "./types";
+import { searchKeywordInCommandGroups } from "./utils";
+>>>>>>> 6c89e1bc02a5e20cf24fdc9edd3f2a642a2b0031
 
 export default function CommandSearch() {
   const commandGroups = commandsRaw as CommandGroup[];
@@ -19,6 +24,7 @@ export default function CommandSearch() {
   const [filteredList, setFilteredList] = useState(commandGroups);
 
   useEffect(() => {
+<<<<<<< HEAD
     const filteredCommandGroups: CommandGroup[] = [];
 
     const notMatchedCommandGroups: CommandGroup[] = [];
@@ -44,13 +50,20 @@ export default function CommandSearch() {
       filteredCommandGroups.push({ key: group.key, commands: list });
     });
     setFilteredList(filteredCommandGroups);
+=======
+    setFilteredList(searchKeywordInCommandGroups(searchText, commandGroups));
+>>>>>>> 6c89e1bc02a5e20cf24fdc9edd3f2a642a2b0031
   }, [searchText]);
 
   return (
     <List
       enableFiltering={false}
       onSearchTextChange={setSearchText}
+<<<<<<< HEAD
       navigationTitle="Search vim commands"
+=======
+      navigationTitle="Search Vim Commands"
+>>>>>>> 6c89e1bc02a5e20cf24fdc9edd3f2a642a2b0031
       searchBarPlaceholder="Learn new command by searching it here."
     >
       {filteredList.map((commandGroup: CommandGroup) => {
@@ -62,8 +75,13 @@ export default function CommandSearch() {
               return (
                 <List.Item
                   key={command.kbd}
+<<<<<<< HEAD
                   title={command.text[0].toUpperCase() + command.text.slice(1)}
                   subtitle={command.kbd}
+=======
+                  title={command.kbd}
+                  subtitle={command.text[0].toUpperCase() + command.text.slice(1)}
+>>>>>>> 6c89e1bc02a5e20cf24fdc9edd3f2a642a2b0031
                 />
               );
             })}