+++ extensions/vim-bro/CHANGELOG.md
@@ -1,7 +1,21 @@
 # Vim-bro Changelog
 
+<<<<<<< HEAD
 ## [Added support to search vim commands] - 2022-03-30
 - Better keyword search
 
 ## [Added vim-bro] - 2022-03-20
 Initial version code
+=======
+## [Added better search with multiple keywords] - 2022-03-30
+
+- Vim-bro can now shows search based on priority defined by number of keyword match.
+
+## [Added support to search vim commands] - 2022-03-30
+
+- Better keyword search
+
+## [Added vim-bro] - 2022-03-20
+
+Initial version code
+>>>>>>> 6c89e1bc02a5e20cf24fdc9edd3f2a642a2b0031